+++ opentelemetry-proto/src/transform/logs.rs
@@ -12,7 +12,7 @@ pub mod tonic {
         transform::common::{to_nanos, tonic::ResourceAttributesWithSchema},
     };
     use opentelemetry::logs::{AnyValue as LogsAnyValue, Severity};
-    use opentelemetry_sdk::export::logs::LogBatch;
+    use opentelemetry_sdk::logs::LogBatch;
     use std::borrow::Cow;
     use std::collections::HashMap;
 
@@ -221,15 +221,39 @@ pub mod tonic {
 mod tests {
     use crate::transform::common::tonic::ResourceAttributesWithSchema;
     use opentelemetry::logs::LogRecord as _;
+    use opentelemetry::logs::Logger as _;
+    use opentelemetry::logs::LoggerProvider as _;
     use opentelemetry::InstrumentationScope;
-    use opentelemetry_sdk::{export::logs::LogBatch, logs::LogRecord, Resource};
+    use opentelemetry_sdk::logs::LogProcessor;
+    use opentelemetry_sdk::logs::{LogResult, LoggerProvider};
+    use opentelemetry_sdk::{logs::LogBatch, logs::LogRecord, Resource};
     use std::time::SystemTime;
+
+    #[derive(Debug)]
+    struct MockProcessor;
+
+    impl LogProcessor for MockProcessor {
+        fn emit(&self, _record: &mut LogRecord, _instrumentation: &InstrumentationScope) {}
+
+        fn force_flush(&self) -> LogResult<()> {
+            Ok(())
+        }
+
+        fn shutdown(&self) -> LogResult<()> {
+            Ok(())
+        }
+    }
 
     fn create_test_log_data(
         instrumentation_name: &str,
         _message: &str,
     ) -> (LogRecord, InstrumentationScope) {
-        let mut logrecord = LogRecord::default();
+        let processor = MockProcessor {};
+        let logger = LoggerProvider::builder()
+            .with_log_processor(processor)
+            .build()
+            .logger("test");
+        let mut logrecord = logger.create_log_record();
         logrecord.set_timestamp(SystemTime::now());
         logrecord.set_observed_timestamp(SystemTime::now());
         let instrumentation =